🌱 폴더명, 파일명 수정
