+++ src/app.module.ts
@@ -5,14 +5,14 @@ import {TypeOrmModule} from "@nestjs/typeorm";
 import * as ormconfig from "../ormconfig";
 import {AppService} from "./app.service";
 import {GatewayModule} from "./chat/chat.module";
-import {UserModule} from "./user/user.module";
-import {HobbyModule} from "./hobbies/hobby.module";
-import {MoodModule} from "./mood/mood.module";
-import {ThemeModule} from "./theme/theme.module";
-import {UserHobbyModule} from "./user-hobby/user-hobby.module";
-import {PaymentModule} from "./payment/payment.module";
-import {OrderController} from "./order/order.controller";
-import {OrderModule} from "./order/order.module";
+import {UserModule} from "./users/user.module";
+import {HobbyModule} from "./hobbies/hobbies.module";
+import {MoodModule} from "./moods/moods.module";
+import {ThemeModule} from "./themes/themes.module";
+import {UserHobbyModule} from "./user-hobbies/user-hobbies.module";
+import {PaymentModule} from "./payments/payments.module";
+import {OrderController} from "./orders/orders.controller";
+import {OrderModule} from "./orders/orders.module";
 
 @Module({
   imports: [
