+++ src/moods/moods.entity.ts
