+++ src/payments/payments.module.ts
@@ -1,10 +1,10 @@
 import {Module} from "@nestjs/common";
-import {PaymentController} from "./payment.controller";
-import {PaymentService} from "./payment.service";
+import {PaymentController} from "./payments.controller";
+import {PaymentService} from "./payments.service";
 import {TypeOrmModule} from "@nestjs/typeorm";
-import {Payment} from "./payment.entity";
-import {Order} from "src/order/order.entity";
-import {User} from "src/user/user.entity";
+import {Payment} from "./payments.entity";
+import {Order} from "src/orders/orders.entity";
+import {User} from "src/users/user.entity";
 
 @Module({
   imports: [TypeOrmModule.forFeature([Payment, Order, User])],
