+++ src/user-hobbies/user-hobbies.module.ts
@@ -1,9 +1,9 @@
 import {Module} from "@nestjs/common";
-import {UserHobbyController} from "./user-hobby.controller";
-import {UserHobbyService} from "./user-hobby.service";
+import {UserHobbyController} from "./user-hobbies.controller";
+import {UserHobbyService} from "./user-hobbies.service";
 import {TypeOrmModule} from "@nestjs/typeorm";
-import {UserHobby} from "./user-hobby.entity";
-import {UserModule} from "../user/user.module";
+import {UserHobby} from "./user-hobbies.entity";
+import {UserModule} from "../users/user.module";
 
 @Module({
   imports: [TypeOrmModule.forFeature([UserHobby]), UserModule],
