+++ src/user-hobbies/user-hobbies.service.ts
@@ -1,8 +1,8 @@
 import {Injectable, NotFoundException} from "@nestjs/common";
 import {InjectRepository} from "@nestjs/typeorm";
 import {Repository} from "typeorm";
-import {User} from "../user/user.entity";
-import {UserHobby} from "./user-hobby.entity";
+import {User} from "../users/user.entity";
+import {UserHobby} from "./user-hobbies.entity";
 
 @Injectable()
 export class UserHobbyService {
